changes made to gofishagame

+++ src/test/java/io/zipcoder/casino/PlayerTest/GoFishPlayerTest.java
@@ -12,14 +12,14 @@ public class GoFishPlayerTest {
     public void constructorTest() {
         String expectedName = "Bob";
         Integer expectedAge = 24;
-        Double expectedBalance = 50.00;
+        Integer expectedBalance = 50;
 
         Player player1 = new Player(expectedName, expectedAge, expectedBalance);
         Player goFishPlayer1 = new GoFishPlayer(player1);
 
         String actualName = goFishPlayer1.getName();
         Integer actualAge = player1.getAge();
-        Double actualBalance = player1.getBalance();
+        Integer actualBalance = player1.getBalance();
 
         Assert.assertEquals(expectedName, actualName);
         Assert.assertEquals(expectedAge, actualAge);
